+++ CHANGELOG.md
@@ -1,7 +1,9 @@
 
 # Change Log
  
-## [Unreleased] - yyyy-mm-dd
+## [Unreleased] - 2025-09-15
+
+- [Hotfix] Fixed bug where the Franka Hand of the follower could not be commanded
  
 ## [0.1.0] - 2025-08-18